+++ modules/deal_steal/includes/classes/CityManager.php
@@ -29,9 +29,8 @@ class CityManager
                 array_push($body, array(
                     $key,
                     $value,
-                    "<a class='icon_delete' title='Delete this city' href='" . SERVER_URL . "modules/deal_steal/control/city_delete.php?city_id=" .
-                        $key . "&module_code=" . $_REQUEST['module_code'] . "'
-                     onclick='return confirmDeletion()'></a>
+                    "<a class='icon_delete confirm_delete' title='Delete this city' href='" . SERVER_URL . "modules/deal_steal/control/city_delete.php?city_id=" .
+                        $key . "&module_code=" . $_REQUEST['module_code'] . "'></a>
                      <a class='icon_edit' title='Update City' href='" . SERVER_URL . "admin/main.php?view=city_update&city_id=" .
                         $key . "&module_code=" . $_REQUEST['module_code'] . "' ></a>"
                 ));
